==> R/2.manuscript.Rmd <==
---
title: "Manuscript"
author: "Kalyn Dorheim"
date: "10/25/2022"
output: 
  html_document:
    toc: yes
    toc_depth: '4'
    toc_float: yes
    number_sections: true
    date: "`r format(Sys.time(), '%d %B, %Y')`"
---

# Obejective 

Visualize and crunch the numbers used to generate figures for the Hector V3 manuscript. 

# Set Up 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, error = FALSE, message = FALSE)
# see https://bookdown.org/yihui/rmarkdown-cookbook/ for more info on markdowns
```

```{r, warning=FALSE, message = FALSE, echo=TRUE}
# Load some custom helpful functions
source(here::here("R", "0.set_up.R"))
source(here::here("R", "0.functions.R"))

# Set up the relative paths
FIGS_DIR  <- file.path(BASE_DIR, "output", "figures")
INPUT_DIR <- file.path(BASE_DIR, "data")
HECTOR_RSLTS_DIR <- file.path(BASE_DIR, "output", "hector_output")

# Define some overall figure aesthetics
THEME_BASE_SIZE <- 10
theme_set(theme_bw(base_size = THEME_BASE_SIZE) + 
            theme(legend.title = element_blank()))
FULL_FIG <- data.frame(width = 7, height = 7)
MANUSCRIPT_FIG <- data.frame(width = 4, height = 3)

COLOR_SCHEME <- c("grey" = "#999999", "col1" = "#56B4E9", "black" = "#000000",
                  "#009E73", "#E69F00", "#F0E442", "#0072B2", "#D55E00", "#CC79A7")
SSP_COLORS <- c("ssp119" = "#00a9cf", "ssp126" = "#003466", "ssp245" = "#f69320",
                "ssp370" = "#df0000", "ssp434" = "#2274ae","ssp460" = "#b0724e",
                "ssp585"= "#980002", "historical" = "#000000", "historical"="#92397a")
OBS_SCHEME <- c("obs" = "grey", "Hector v3" = "#E69F00")

hector_version <- HECTOR_VERSION
```

## Load Hector Results

```{r, echo=TRUE}
list.files(HECTOR_RSLTS_DIR, pattern = "hector_3.2.0_ssp.csv", full.names = TRUE) %>%
  lapply(read.csv) %>%
  do.call(what = "rbind") %>% 
  filter(year <= 2100) -> 
  hector_ssp

list.files(HECTOR_RSLTS_DIR, pattern = "hector_3.2.0_ssp-conc.csv", full.names = TRUE) %>%
  lapply(read.csv) %>%
  do.call(what = "rbind") %>% 
  filter(year <= 2100) -> 
  hector_ssp_conc
```



# Comparison with Observations 

## Atmospheric CO~2~ 


I've added the uncertainty from Keeling record to the Meinshausen product and you'll notice that is is so small it doesn't really add of an uncerintaty cone. 

```{r}
# Meinshausen, M., Vogel, E., Nauels, A., Lorbacher, K., Meinshausen, N., Etheridge, D. M., 
# Fraser, P. J., Montzka, S. A., Rayner, P. J., Trudinger, C. M., Krummel, P. B., Beyerle, 
# U., Canadell, J. G., Daniel, J. S., Enting, I. G., Law, R. M., Lunder, C. R., O'Doherty, 
# S., Prinn, R. G., Reimann, S., Rubino, M., Velders, G. J. M., Vollmer, M. K., Wang, 
# R. H. J., and Weiss, R.: Historical greenhouse gas concentrations for climate modelling 
# (CMIP6), Geosci. Model Dev., 10, 2057–2116, https://doi.org/10.5194/gmd-10-2057-2017, 2017.
here::here(INPUT_DIR, "Supplementary_Table_UoM_GHGConcentrations-1-1-0_annualmeans_v23March2017.csv") %>%  
  read.csv(skip = 21) %>% 
  na.omit %>% 
  select(year =  "v.YEARS.GAS..", value = "CO2") %>% 
  mutate(variable = "co2 obs") -> 
  cmip6_co2_obs

hector_ssp %>% 
  filter(version == hector_version & variable == CONCENTRATIONS_CO2()) %>% 
  filter(scenario == "ssp119") %>%
  filter(year <= 2014) -> 
  hector_co2

# The cmip6_co2_obs data needs to be split into two categories to indicate the period of time 
# that was used in the calibration process. 
cmip6_co2_obs %>% 
  filter(year <= 1800) -> 
  inital_obs


cmip6_co2_obs %>% 
  filter(year >= 1800) -> 
  cmip6_co2_obs

ggplot() + 
  geom_point(data = inital_obs, aes(year, value, color = "Meinshausen et al 2017")) +
  geom_line(data = cmip6_co2_obs, aes(year, value, color = "Meinshausen et al 2017")) +
  geom_line(data = hector_co2, aes(year, value, color = "Hector V3")) + 
  scale_color_manual(values = c("Hector V3" = OBS_SCHEME[["Hector v3"]],
                                "Meinshausen et al 2017" = "black")) + 
  scale_fill_manual(values = c("Meinshausen et al 2017" = "black")) + 
  labs(x = "Year", y = expression('CO'[2]~' (ppm)')) + 
  theme(legend.position = "bottom") -> 
  plot; plot

ggsave(plot, filename = file.path(FIGS_DIR, "Fig-2.png"), 
       width = MANUSCRIPT_FIG$width, height = MANUSCRIPT_FIG$height)
```

Evaluate historical performance of the model. 

```{r}
# Manipulate the data frames so that it is easy to calculate the MSE. 
cmip6_co2_obs %>% 
  select(year, co2_obs = value) -> 
  cmip6_co2_obs_wide

# Calculate the RMSE over the entire observational record. 
hector_co2 %>% 
  filter(year %in% cmip6_co2_obs$year) %>% 
  select(year, hector_co2 = value) %>% 
  inner_join(cmip6_co2_obs_wide) -> 
  wide_data
```


RMSE between Hector and observations. 

```{r}
# Calculate the RMSE 
wide_data %>% 
  mutate(SE = (hector_co2-co2_obs)^2) %>% 
  pull(SE) %>% 
  mean() %>% 
  sqrt() -> 
  RMSE_co2
RMSE_co2
```



Linear regression  

```{r}
wide_data %>% 
  ggplot(aes(co2_obs, hector_co2)) + 
  geom_point() + 
  labs(x = expression('Meinshausen et al 2017 CO'[2]~' (ppm)'),  
       y = expression('Hector CO'[2]~' (ppm)'))

fit <- lm(data = wide_data, formula = co2_obs~hector_co2)
summary(fit)
```


## Global Mean Surface Temperature 

```{r}
# Hadcrut5
# Global mean surface temperature anomaly
# https://www.metoffice.gov.uk/hadobs/hadcrut5/data/current/download.html
# The temperature anomaly is based off of 1961–1990
# surface temperature with the anomaly!
# Morice, C. P., Kennedy, J. J., Rayner, N. A., Winn, J. P., Hogan, E., Killick, R. E., et al. (2021).
# An updated assessment of near-surface temperature change from 1850: the HadCRUT5 data
# set. Journal of Geophysical Research: Atmospheres, 126, e2019JD032361.
# https://doi.org/10.1029/2019JD032361
here::here(INPUT_DIR, "HadCRUT5.csv") %>%
  read.csv(stringsAsFactors = FALSE) %>% 
  na.omit ->
  hadcrut_obs_data
names(hadcrut_obs_data) <- c("year", "value", "lower", "upper")
hadcrut_obs_data$variable <- "gmst"
hadcrut_obs_data

```


Mean and SD for the GMST observations 

```{r}
mean(hadcrut_obs_data$upper - hadcrut_obs_data$lower)
sd(hadcrut_obs_data$value)
```

```{r}
yrs <- hadcrut_obs_data$year

hector_ssp %>% 
  filter(version == hector_version & variable == "gmst") %>% 
  filter(scenario == "ssp119") %>%
  filter(year %in% yrs) %>% 
  normalize_to_hadcrut %>% 
  na.omit -> 
  hector_gmst

ggplot() + 
  geom_ribbon(data = hadcrut_obs_data, aes(year, ymin = lower, ymax = upper), alpha = 0.4) + 
  geom_line(data = hadcrut_obs_data, aes(year, value, color = "Morice et al., 2021")) + 
  geom_line(data = hector_gmst, aes(year, value, color = "Hector V3"), linewidth = 0.75) + 
  scale_color_manual(values = c("Hector V3" = OBS_SCHEME[["Hector v3"]],
                                "Morice et al., 2021" = "black")) + 
  labs(x = "Year", y = expression("Temperature anomaly ("~degree~"C)")) + 
  theme(legend.position = "bottom") -> 
  plot; plot
big_plot <- plot 

ggsave(plot, filename = file.path(FIGS_DIR, "Fig-3.png"), 
       width = MANUSCRIPT_FIG$width+1, height = MANUSCRIPT_FIG$height + 1)
```

Make the inset figure. 

```{r}
hector_gmst_decade <- data.frame(value =  rollmean(hector_gmst$value, 10), 
                                 sd = rollapply(hector_gmst$value, width = 10, FUN=sd), 
                                 year = hector_gmst$year[5:167])

hadcrut_obs_data_decade <- data.frame(value = rollmean(hadcrut_obs_data$value, 10), 
                                      sd = rollapply(hadcrut_obs_data$value, width = 10, FUN=sd),
                                      year = hadcrut_obs_data$year[5:167])


ggplot() + 
  geom_line(data = hadcrut_obs_data_decade, aes(year, value, color = "Morice et al., 2021")) + 
  geom_ribbon(data = hadcrut_obs_data_decade, aes(year, ymin = value - abs(sd), ymax = value + abs(sd), 
                                                  fill = "Morice et al., 2021"), alpha = 0.2) + 
  geom_line(data = hector_gmst_decade, aes(year, value, color = "Hector V3"), linewidth = 0.75) + 
  geom_ribbon(data = hector_gmst_decade, aes(year, ymin = value - abs(sd), ymax = value + abs(sd), 
                                             fill = "Hector V3"), alpha = 0.4) + 
  scale_color_manual(values = c("Hector V3" = OBS_SCHEME[["Hector v3"]],
                                "Morice et al., 2021" = "black")) + 
  scale_fill_manual(values = c("Hector V3" = OBS_SCHEME[["Hector v3"]],
                               "Morice et al., 2021" = "black")) + 
  labs(x = NULL, y = NULL) + 
  theme(legend.position = "none") -> 
  mini_plot


ggsave(mini_plot, filename = file.path(FIGS_DIR, "Fig-3_inset.png"), 
       width = MANUSCRIPT_FIG$width+1, height = MANUSCRIPT_FIG$height + 1)

```


TODO add code to add the inset figure to the bigger figure. 

Calculate RMSE and model-observation fit

```{r}
# Manipulate the data to make the RMSE comparison easy to do. 
hadcrut_obs_data %>% 
  select(year, obs = value) -> 
  wide_gmst_obs

hector_gmst %>% 
  select(year, hector_gmst = value) %>% 
  inner_join(wide_gmst_obs) -> 
  gmst_wide

gmst_wide %>% 
  mutate(SE = (hector_gmst - obs)^2) %>%  
  pull(SE) %>% 
  mean() %>% 
  sqrt() -> 
  RMSE 
RMSE
```

Observation - Model fit

```{r}
fit <- lm(data = gmst_wide, hector_gmst~obs)
summary(fit)

gmst_wide %>% 
  ggplot(aes(obs, hector_gmst)) + 
  geom_point() + 
  labs(x = expression('Morice et al 2020 GMST ('~degree~"C)"),  
       y = expression('Hector GMST ('~degree~"C)"))
```
```{r}
hector_gmst %>% 
  filter(year %in% 2012:2021) %>% 
  pull(value) %>% mean()

hector_gmst %>% 
  filter(year %in% 2012:2021) %>% 
  pull(value) %>% sd()
```



# CMIP6 Comparison 

Make a plot with the ESMs colored by their ECS labels. 


```{r}
scns <- c("ssp245", "ssp126", "ssp585")

# G. A. Meehl, C. A. Senior, V. Eyring, G. Flato, J.-F. Lamarque, R. J. Stouffer, K. E. Taylor,
# M. Schlund, Context for interpreting equilibrium climate sensitivity and transient climate response
# from the CMIP6 Earth system models. Sci. Adv. 6, eaba1981 (2020). 
# Schlund, Manuel, Axel Lauer, Pierre Gentine, Steven C. Sherwood, and Veronika Eyring. 2020. 
# “Emergent Constraints on Equilibrium Climate Sensitivity in CMIP5: Do They Hold for CMIP6?” Earth System Dynamics 11 (4): 1233–58.

# Lovato, T., Peano, D., Butenschön, M., Materia, S., Iovino, D., Scoccimarro, E.,
# et al. (2022). CMIP6 simulations with the CMCC Earth System Model (CMCCESM2). 
# Journal of Advances in Modeling Earth Systems, 14, e2021MS002814.
# https://doi.org/10.1029/2021MS002814
cmip6_ecs <- as.data.frame(rbind(c("ACCESS-CM2", 4.7),
                                 c("ACCESS-ESM1-5", 3.9),
                                 c("CAMS-CSM1-0", 2.3),
                                 c("CanESM5", 5.6),
                                 c("CESM2", 5.2),
                                 c("CESM2-WACCM", 4.8),
                                 c("CMCC-CM2-SR5", 3.52) ,  
                                 c("HadGEM3-GC31-LL", 5.6),
                                 c("MIROC-ES2L", 2.7),
                                 c("MIROC6", 2.6),
                                 c("MRI-ESM2-0", 3.2),
                                 c("NorESM2-MM", 2.5),
                                 c("TaiESM1", 4.31),        
                                 c("UKESM1-0-LL", 5.3), 
                                 c("CMCC-ESM2", 3.57)))
names(cmip6_ecs) <- c("model", "ecs")


# TS.3.2 Climate Sensitivity and Earth System Feedbacks
# the likely range is 2.5°C to 4°C and the very likely range is 2°C to 5°C. 
cmip6_ecs %>%  
  mutate(id = "not very likely") %>% 
  mutate(id = ifelse(ecs >= 2 & ecs <= 5, "very likely", id)) -> 
  ecs_table

read.csv(here::here("data", "cmip6_model_means.csv")) %>% 
  dplyr::filter(scenario %in% scns)  %>% 
  filter(model %in% cmip6_ecs$model) %>% 
  full_join(ecs_table, by = "model") -> 
  cmip6_rslts 

hector_ssp_conc %>% 
  filter(year %in% 1850:2100) %>% 
  filter(version == hector_version) %>% 
  filter(variable %in% c(SST(), GLOBAL_TAS(),  LAND_TAS())) %>% 
  filter(scenario %in% cmip6_rslts$scenario) -> 
  hector_temp

# Normalize the Hector temperatures to the CMIP6 reference period
split(hector_temp, interaction(hector_temp$variable, hector_temp$scenario), drop = TRUE) %>% 
  lapply(FUN = normalize_hector_temp, period = 1850:1900) %>%  
  do.call(what = "rbind") -> 
  hector_temp

cmip6_rslts %>% 
  group_by(scenario, year, variable, id) %>% 
  summarise(min = min(value), 
            max = max(value)) -> 
  cmip6_temp_summary

# Create the labels 
temp_labs <- c("Global Mean Air Temp.", "Mean Land Surface Temp.", "Mean Sea Surface Temp.")
names(temp_labs) <-   c(GLOBAL_TAS(), LAND_TAS(), SST())

ggplot() + 
  geom_ribbon(data = cmip6_temp_summary, aes(year, ymin = min, ymax = max, fill = id), alpha = 0.9) + 
  geom_line(data = hector_temp, aes(year, value, color = "Hector V3"), size = 1) + 
  facet_grid(scenario~variable, labeller = labeller(variable = temp_labs), scales = "free") + 
  labs(y = expression("Temperature anomaly relative to 1850-1860 ("~degree~"C)"), x = "Year") + 
  scale_color_manual(values = c("Hector V3" = OBS_SCHEME[["Hector v3"]])) + 
  scale_fill_manual(values = c("very likely" = "#5A5A5A", 
                               "not very likely" = "#D3D3D3")) + 
  theme(panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(), 
        legend.position = "bottom") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=1)) -> 
  plot; plot

# ggsave(plot, filename = file.path(FIGS_DIR, "hector_cmip_temps-ranges.png"), 
#        width = MANUSCRIPT_FIG$width * 1.25, height = MANUSCRIPT_FIG$height * 1.25)
```

```{r}

ggplot() + 
  geom_line(data = cmip6_rslts, aes(year, value, group = interaction(model), color = "CMIP6 ESM"), alpha = 0.5) + 
  geom_line(data = hector_temp, aes(year, value, color = "Hector V3"),) + 
  facet_grid(scenario~variable, labeller = labeller(variable = temp_labs), scales = "free") + 
  labs(y = expression("Temperature anomaly ("~degree~"C)"), x = "Years") + 
  scale_color_manual(values = c("Hector V3" = OBS_SCHEME[["Hector v3"]], 
                                "CMIP6 ESM" = "#5A5A5A")) + 
  theme(panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(), 
        legend.position = "bottom") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) -> 
  plot; plot


ggsave(plot, filename = file.path(FIGS_DIR, "Fig-4.png"), 
       width = MANUSCRIPT_FIG$width * 1.25, height = MANUSCRIPT_FIG$height * 1.25)

```


## Idealized Runs 


```{r}
max_yr <- 150
hector_idealized <- read.csv(file.path(HECTOR_RSLTS_DIR, "hector_3.2.0_idealized.csv")) %>% 
  filter(variable == GLOBAL_TAS()) %>% 
  mutate(scenario = if_else(scenario ==  "abruptx4CO2", "abrupt-4xCO2", scenario)) %>% 
  filter(year <= max_yr)
cmip6_idealized <- read.csv(here::here("data", "cmip6_idealized.csv")) %>% 
  filter(year <= max_yr) %>% 
  rename(scenario = experiment)


ggplot() + 
  geom_line(data = cmip6_idealized, aes(year, value, group = interaction(model, ensemble),
                                        color = "CMIP6 ESM"), alpha = 0.5) + 
  geom_line(data = hector_idealized, aes(year, value, color = "Hector V3"), 
            linewidth = 1) + 
  facet_wrap("scenario", scales = "free") + 
  labs(y = expression("Temperature anomaly ("~degree~"C)"), x = "Years") + 
  scale_color_manual(values = c("Hector V3" = OBS_SCHEME[["Hector v3"]],
                                "CMIP6 ESM" = "#5A5A5A",
                                "rcmip" = "black")) +
  theme(panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(), 
        legend.position = "bottom") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) -> 
  plot; plot


ggsave(plot, filename = file.path(FIGS_DIR, "Fig-5.png"), 
       width = MANUSCRIPT_FIG$width * 1.25, height = MANUSCRIPT_FIG$height * 1.25)

```


# IPCC AR6 comparisons 

## Recreating IPCC SPM10 

```{r, warning=FALSE}
# Read in the IPCC results 
ipcc_data <- read.csv(here::here("data", "IPCC_co2_tas.csv"))
hector_tas_v_co2 <- read.csv(here::here("output", "hector_output", "hector_3.2.0_IPCCemiss_co2_tas.csv"))

colors_to_use <- SSP_COLORS[names(SSP_COLORS) %in% ipcc_data$scenario]

ggplot() +
  geom_ribbon(data = ipcc_data, aes(co2, ymin = min, 
                                    ymax = max, fill = scenario), alpha = 0.1) + 
  geom_line(data = ipcc_data, aes(co2, obs, color = "Obs", linetype = "Observations")) +
  geom_line(data = hector_tas_v_co2, aes(co2, tas, color = scenario), linewidth = 0.75) +
  scale_fill_manual(values = colors_to_use) + 
  scale_color_manual(values = colors_to_use) + 
  labs(y = expression("Temperature anomaly relative to 1850-1900 ("~degree~"C)"),
       x = expression("Cumlative "~CO[2]~" Emissions")) + 
  theme_bw(base_size = THEME_BASE_SIZE) +
  theme(plot.background = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  scale_linetype_manual(values = c("Observations" = 2)) +
  labs(color = "Hector", fill = "IPCC AR6", linetype = element_blank()) + 
  NULL -> 
  plot; plot

# ggsave(plot, filename = file.path(FIGS_DIR, "recreate_IPCC_fig.png"), 
#        width = MANUSCRIPT_FIG$width * 2, height = MANUSCRIPT_FIG$height * 1.75)
```


## Recreating Table SPM.1

```{r}
hector_ssp %>% 
  filter(variable == "gmst") %>% 
  filter(scenario == "ssp119") %>% 
  filter(year %in% 1850:1900) %>% 
  pull(value) %>% 
  mean -> 
  ref

hector_ssp %>% 
  filter(!grepl(pattern = "conc", x = scenario)) %>%  
  filter(scenario %in% c("ssp119", "ssp126", "ssp245", "ssp370", "ssp585")) %>% 
  filter(version == hector_version) %>% 
  filter(variable == "gmst") %>% 
  mutate(value = value - ref) %>% 
  mutate(period = "x") %>% 
  mutate(period = if_else(year %in% 2021:2040, "near term", period)) %>% 
  mutate(period = if_else(year %in% 2041:2060, "mid term", period)) %>% 
  mutate(period = if_else(year %in% 2081:2100, "long term", period)) %>% 
  group_by(scenario, period) %>% 
  summarise(value = mean(value)) %>% 
  spread(period, value) %>% 
  select(-x)
```


# Hector's Post Spin Up State 

```{r}
# The post spin up C cycle values come from the ocean.log/simpelN.log file and the 
# the ini files. 
variables = c(VEG_C(), 
              DETRITUS_C(), 
              SOIL_C(), 
              PERMAFROST_C(), 
              CONCENTRATIONS_CO2(), 
              OCEAN_PREIND_C_ID(), 
              OCEAN_PREIND_C_SURF())
post_spinup = 
  c(562, 62.3489, 2030.59, 865, 277.15, 27118.3 + 8784.06, 146.61 + 818.131)
inital_pools = c(550, 55, 1782, 865, 277.15, 37100, 900)

data.frame(variables, inital_pools, post_spinup) %>% 
  mutate(dif = inital_pools - post_spinup) %>% 
  write.csv(file.path(HECTOR_RSLTS_DIR, "inital_carbon_pools.csv"), row.names = FALSE)
```


# Hector's contemporary state (e.g. magnitude of land and ocean sinks, ocean pH, NPP, etc)

## GCP Land Sink

The land CO2 sink continued to increase during the 2012–2021 period primarily in response to increased atmospheric CO2, albeit with large interannual variability. The land CO2 sink was 3.1 ± 0.6 GtC yr−1 during the decade 2012–2021 (29 % of total CO2 emissions), 0.4 GtC yr−1 larger than during the previous decade (2000–2009), with a preliminary 2022 estimate of around 3.4 GtC yr−1.^[https://essd.copernicus.org/articles/14/4811/2022/]

```{r}
hector_ssp %>% 
  filter(!grepl(pattern = "conc", x = scenario)) %>%  
  filter(variable == "land sink") %>% 
  filter(version == hector_version) %>% 
  filter(scenario == "ssp245") %>%
  filter(year %in% 2012:2021) %>% 
  group_by(variable) %>% 
  summarise(mean = mean(value), 
            sd = sd(value)) -> 
  modern_decade; modern_decade
```
## GCP Ocean Sink 

The ocean CO2 sink was 2.9 ± 0.4 GtC yr−1 during the decade 2012–2021 (26 % of total CO2 emissions), with a similar preliminary estimate of 2.9 GtC yr−1 for 2022.^[https://essd.copernicus.org/articles/14/4811/2022/]

```{r}
hector_ssp %>% 
  filter(!grepl(pattern = "conc", x = scenario)) %>%  
  filter(variable == OCEAN_UPTAKE()) %>% 
  filter(year %in% 2012:2021) %>% 
  group_by(variable, units) %>% 
  summarise(mean = mean(value), 
            sd = sd(value)) 
```

# ocean pH, & NPP

```{r}
hector_ssp %>% 
  filter(!grepl(pattern = "conc", x = scenario)) %>%  
  filter(variable == PH()) %>% 
  filter(year %in% 2012:2021) %>% 
  group_by(variable, units) %>% 
  summarise(mean = mean(value), 
            sd = sd(value)) 
```

```{r}
hector_ssp %>% 
  filter(!grepl(pattern = "conc", x = scenario)) %>%  
  filter(variable == PH()) %>% 
  filter(year %in% 2012:2021) 
```

```{r}
hector_ssp %>% 
  filter(variable %in% c("land sink", "gmst", "ocean_uptake", "pH", "sst", "NPP")) %>% 
  filter(scenario == "ssp245") %>%
  filter(year %in% 2012:2021) %>% 
  group_by(variable, units) %>% 
  summarise(mean = mean(value), 
            sd = sd(value))
```
